requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
class-registry==2.1.2
filters==1.3.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
numpy==1.24.2
Pillow==9.4.0
python-dateutil==2.8.2
python-telegram-bot==20.1
pytz==2022.7.1
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
style==1.1.0
telegram==0.0.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchvision==0.14.1+cu117
typing-extensions==4.5.0
update==0.0.1
urllib3==1.26.14
